Improve presentation style for better readability
Change system colors to have a mostly black and white theme with
a focus on text and content.

Update text to have more readable text that preserves the case
of code, styling code blocks and inline code like GitHub.

Removes image borders and background styles so images with
transparant backgrounds fit much better in the presentation.