Port presntation to flexible reveal framework
Supports splitting each slide into its own file; this significantly
simplifies the process of updating and adding content to the
presentation.

Previously:
- changes which affected slide order were hard to follow
- changes which changed slide order and content for a related
  section were hard to follow
- updating the theme was very manual
- manually editing html files was very tedious

Now:
- support for either html or markdown, whichever works for a
  particular situation
- allows changes which affect slide order to be totally separate
  from slide content
- allows changes which change slide attributes to be totally
  separate from changes which affect slide content
- much easier to write slide notes

The reveal generator does, however, make it more difficult to
keep a copy of the slides available online, and will require a
separate GitHub repository to be kept online.